Move the stats tests into their own file
Adding the C-01 and C-02 regression tests pushed stats.rs to 311 lines,
over the 300-line cap. Split to stats/tests.rs, matching the pattern
netscli-mcp already uses for dispatch/tests.rs.

Verified no tests were lost rather than assuming it: origin/main runs 65
lib tests, this branch runs 67 -- exactly the two added.
